Refactor into ufacter
This takes original work (thanks) and creates new util called ufacter.
It's essentially small facter implementation in go (it uses goroutines
now) but the main goal is to provide extra (new) facts about network via
netlink (low level link information, relationships, routing info). These
things are missing in PuppetLab's facter and they can't be easily
written in pure Ruby (facter 4.x will be again Ruby only) as netlink
binding liraries are sparse and not finished. It's easier to call an
external app via custom fact:

    Facter.add(:link) do
      setcode do
        json = `ufacter -json -modules link`
        JSON.parse(json)['link']
      end
    end

The idea is to keep the ufactor as a "loose drop-in" replacement of the
original facter because it's cool idea and it might find some use in our
project Foreman, especially Foreman Discovery Image.